Project: Latesummer Description:封装分页查询（包含搜索条件）

+++ src/main/java/com/latesummer/controller/UserController.java
@@ -21,7 +21,7 @@ public class UserController {
 	private IUserService userService;
 
 	@RequestMapping(value = "/add/{id}/{name}/{age}")
-	public User addUser(@PathVariable Long id, @PathVariable String name, @PathVariable Integer age) {
+	public User addUser(@PathVariable Integer id, @PathVariable String name, @PathVariable Integer age) {
 		User user = new User();
 		user.setId(id);
 		user.setName(name);
